Factors timepoints out of keyframes.

+++ src/animation/runtime/animation_keyframe.h
@@ -47,7 +47,7 @@ namespace animation {
 // Translation values are stored as half precision floats with 16 bits per
 // component.
 struct Float3Key {
-  float ratio;
+  uint16_t ratio;
   uint16_t previous;
   uint16_t value[3];
 };
@@ -68,7 +68,7 @@ struct Float3Key {
 // key frames, but in this case RotationKey structure would induce 16 bits of
 // padding.
 struct QuaternionKey {
-  float ratio;
+  uint16_t ratio;
   uint16_t previous : 13;
   uint16_t largest : 2;  // The largest component of the quaternion.
   uint16_t sign : 1;     // The sign of the largest component. 1 for negative.
